Add files via upload
This is my webportfolio showing my academic achievement and degrees. It also shows my technical skills like programming and computer networking skills and years of experience. It also shows services i provide. The webportfolio has a form for users to fill, if the form isnt filled correctly and completely there will be a prompt to fill the form correctly and completely before submission action to the localStorage can take place.